Score handling moved under scoreHandlingFunctions Updated Tops functionalities for both all games and single game queries to reflect the new db and interface
