+++ SlackAPI/scoreHandlingFunctions.js
